fix(cannon): avoid uint64->int conversion in block preload loop (gosec G115)
The block-preload worker loop converted BlockPreloadWorkers (uint64) to int,
which gosec flags as a potential integer-overflow conversion (G115). The loop
index isn't used as an int, so iterate in uint64 directly and drop the cast.

+++ pkg/cannon/ethereum/beacon.go
@@ -160,7 +160,7 @@ func (b *BeaconNode) Start(ctx context.Context) error {
 
 	go b.blockCache.Start()
 
-	for i := 0; i < int(b.config.Beacon.BlockPreloadWorkers); i++ {
+	for i := uint64(0); i < b.config.Beacon.BlockPreloadWorkers; i++ {
 		go func() {
 			for identifier := range b.blockPreloadChan {
 				b.metrics.SetPreloadBlockQueueSize(string(b.Metadata().Network.Name), len(b.blockPreloadChan))
